+++ functional/promise/memo_test.go
@@ -1,4 +1,4 @@
-package functional
+package promise
 
 import (
 	"context"
